+++ easycv/models/detection/utils/__init__.py
@@ -3,5 +3,4 @@
 
 from .boxes import (bbox2result, bboxes_iou, box_cxcywh_to_xyxy,
                     box_xyxy_to_cxcywh, generalized_box_iou, postprocess, bbox_overlaps, distance2bbox)
-from .misc import (accuracy, interpolate, multi_apply, select_single_mlvl, filter_scores_and_topk)
-from .utils import output_postprocess
+from .misc import (accuracy, interpolate, multi_apply, select_single_mlvl, filter_scores_and_topk, output_postprocess, fp16_clamp)
