+++ easycv/models/detection/utils/misc.py
@@ -7,6 +7,7 @@ import torchvision
 from packaging import version
 from six.moves import map, zip
 from torch import Tensor
+import numpy as np
 
 if version.parse(torchvision.__version__) < version.parse('0.7'):
     from torchvision.ops import _new_empty_tensor
@@ -152,3 +153,43 @@ def filter_scores_and_topk(scores, score_thr, topk, results=None):
             raise NotImplementedError(f'Only supports dict or list or Tensor, '
                                       f'but get {type(results)}.')
     return scores, labels, keep_idxs, filtered_results
+
+def output_postprocess(outputs, img_metas=None):
+    detection_boxes = []
+    detection_scores = []
+    detection_classes = []
+    img_metas_list = []
+
+    for i in range(len(outputs)):
+        if img_metas:
+            img_metas_list.append(img_metas[i])
+        if outputs[i] is not None:
+            bboxes = outputs[i][:, 0:4] if outputs[i] is not None else None
+            if img_metas:
+                bboxes /= img_metas[i]['scale_factor'][0]
+            detection_boxes.append(bboxes.cpu().numpy())
+            detection_scores.append(
+                (outputs[i][:, 4] * outputs[i][:, 5]).cpu().numpy())
+            detection_classes.append(outputs[i][:, 6].cpu().numpy().astype(
+                np.int32))
+        else:
+            detection_boxes.append(None)
+            detection_scores.append(None)
+            detection_classes.append(None)
+
+    test_outputs = {
+        'detection_boxes': detection_boxes,
+        'detection_scores': detection_scores,
+        'detection_classes': detection_classes,
+        'img_metas': img_metas_list
+    }
+
+    return test_outputs
+
+
+def fp16_clamp(x, min=None, max=None):
+    if not x.is_cuda and x.dtype == torch.float16:
+        # clamp for cpu float16, tensor fp16 has no clamp implementation
+        return x.float().clamp(min, max).half()
+
+    return x.clamp(min, max)
